tests/scripts/requirements.txt
appdirs==1.4.3
Babel==2.4.0
beautifulsoup4==4.4.1
catfish==1.4.2
chardet==2.3.0
command-not-found==0.3
debtcollector==1.13.0
defer==1.0.6
deprecation==1.0
html5lib==0.999
iso8601==0.1.11
keystoneauth1==2.19.0
language-selector==0.1
lightdm-gtk-greeter-settings==1.2.1
lxml==3.5.0
menulibre==2.1.3
monotonic==1.3
msgpack-python==0.4.8
mugshot==0.3.1
netaddr==0.7.19
netifaces==0.10.5
onboard==1.2.0
openstacksdk==0.9.14
os-client-config==1.26.0
oslo.i18n==3.15.0
oslo.serialization==2.18.0
oslo.utils==3.25.0
pbr==2.1.0
pexpect==4.0.1
Pillow==3.1.2
positional==1.1.1
prettytable==0.7.2
psutil==3.4.2
ptyprocess==0.5
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pyparsing==2.2.0
python-apt==1.1.0b1
python-debian==0.1.27
python-novaclient==8.0.0
python-systemd==231
pytz==2017.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.3.0
requests==2.12.5
requestsexceptions==1.2.0
sessioninstaller==0.0.0
simplejson==3.10.0
six==1.10.0
ssh-import-id==5.5
stevedore==1.21.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
wrapt==1.10.10
xkit==0.0.0
